+++ R/plot_ess.R
@@ -1,11 +1,12 @@
 
-#' @title Density Plot Diagnostics
+#' @title Plot Effective Sample Size Diagnostics
 #' @author Sam Schildhauer
 #' @description
-#'  plot_jags_dens() takes a [list] output from [serodynamics::run_mod()]
-#'  to create density plots for each chain run in the mcmc estimation.
+#'  plot_ess() takes a [list] output from [serodynamics::run_serodynamics()]
+#'  to create summary diagnostics for each chain run in the mcmc estimation.
 #'  Defaults will produce every combination of antigen/antibody, parameters,
-#'  and stratifications, unless otherwise specified.
+#'  and stratifications, unless otherwise specified. At least 2 chains are 
+#'  required to run function. 
 #'  Antigen/antibody combinations and stratifications will vary by analysis.
 #'  The antibody dynamic curve includes the following parameters:
 #'  - y0 = baseline antibody concentration
@@ -13,32 +14,35 @@
 #'  - t1 = time to peak
 #'  - r = shape parameter
 #'  - alpha = decay rate
-#' @param data A [list] outputted from run_mod().
+#' @param data A [list] outputted from [run_serodynamics()].
 #' @param iso Specify [character] string to produce plots of only a
 #' specific antigen/antibody combination, entered with quotes. Default outputs
 #' all antigen/antibody combinations.
 #' @param param Specify [character] string to produce plots of only a
 #' specific parameter, entered with quotes. Options include:
-#' - `alpha` = posterior estimate of decay rate
-#' - `r` = posterior estimate of shape parameter
-#' - `t1` = posterior estimate of time to peak
 #' - `y0` = posterior estimate of baseline antibody concentration
 #' - `y1` = posterior estimate of peak antibody concentration
+#' - `t1` = posterior estimate of time to peak
+#' - `r` = posterior estimate of shape parameter
+#' - `alpha` = posterior estimate of decay rate
 #' @param strat Specify [character] string to produce plots of specific
 #' stratification entered in quotes.
-#' @return A [base::list()] of [ggplot2::ggplot()] objects producing density
-#' plots for all the specified input.
+#' @return A [list] of [ggplot2::ggplot] objects showing the 
+#' proportion of effective samples taken/total samples taken for all parameter
+#' iso combinations. The estimate with the highest proportion of effective
+#' samples taken will be listed first.
 #' @export
-#' @example inst/examples/examples-plot_jags_densitydx.R
+#' @example inst/examples/examples-plot_ess.R
 
-plot_jags_dens <- function(data,
-                           iso = unique(data$Iso_type),
-                           param = unique(data$Parameter),
-                           strat = unique(data$Stratification)) {
+
+plot_ess <- function(data,
+                     iso = unique(data$Iso_type),
+                     param = unique(data$Parameter),
+                     strat = unique(data$Stratification)) {
   
   attributes_jags <- data[["attributes"]]
 
-  dens_strat_list <- list()
+  eff_strat_list <- list()
   for (i in strat) {
 
     visualize_jags_sub <- data |>
@@ -46,7 +50,7 @@ plot_jags_dens <- function(data,
       dplyr::filter(.data$Subject == "newperson")
 
     # Creating open list to store ggplots
-    density_out <- list()
+    eff_out <- list()
     # Looping through the isos
     for (j in iso) {
       visualize_jags_plot <- visualize_jags_sub |>
@@ -59,20 +63,26 @@ plot_jags_dens <- function(data,
 
       visualize_jags_plot <- visualize_jags_plot |>
         # Changing parameter name to reflect the input
-        dplyr::mutate(Parameter = paste0("iso = ", j, ", parameter = ",
-                                         .data$Parameter, ", strat = ",
-                                         i),
-                      value = log(.data$value))
+        dplyr::mutate(Parameter = .data$Parameter)
       # Assigning attributes, which are needed to run ggs_density
       attributes(visualize_jags_plot) <- c(attributes(visualize_jags_plot),
                                            attributes_jags)
+
       # Creating density plot
-      densplot <- ggmcmc::ggs_density(visualize_jags_plot) +
-        ggplot2::theme_bw() +
-        ggplot2::labs(x = "log(value)")
-      density_out[[j]] <- densplot
+      eff <- ggmcmc::ggs_effective(visualize_jags_plot) +
+        ggplot2::theme_bw()  +
+        ggplot2::labs(title = "Effective sample size",
+                      subtitle = plot_title_fun(i, j),
+                      x = "Proportion of effective samples") +
+        ggplot2::scale_y_discrete(limits = c("alpha", "shape", "t1", "y1", 
+                                             "y0"))
+      eff_out[[j]] <- eff
     }
-    dens_strat_list[[i]] <- density_out
+    eff_strat_list[[i]] <- eff_out
   }
-  dens_strat_list
+  #Printing only one plot if only one exists.
+  if (sum(lengths(eff_strat_list)) == 1) {
+    eff_strat_list <- eff_strat_list[[1]][[iso]]
+  } 
+  eff_strat_list
 }