+++ inst/examples/examples-plot_density.R
@@ -1,0 +1,8 @@
+
+data <- serodynamics::nepal_sees_jags_output
+
+# Specifying isotype and stratification for traceplot.
+plot_density(
+             data = data,
+             iso = "HlyE_IgA",
+             strat = "typhi")