+++ inst/examples/examples-plot_ess.R
@@ -1,0 +1,6 @@
+
+data <- serodynamics::nepal_sees_jags_output
+
+plot_ess(data = data,
+         iso = "HlyE_IgA",
+         strat = "typhi")