+++ inst/examples/examples-plot_trace.R
@@ -2,7 +2,7 @@
 data <- serodynamics::nepal_sees_jags_output
 
 # Specifying isotype, parameter, and stratification for traceplot.
-plot_jags_trace(
-                data = data,
-                iso = "HlyE_IgA",
-                strat = "typhi")
+plot_trace(
+           data = data,
+           iso = "HlyE_IgA",
+           strat = "typhi")